wip: initial work for adding users to tenant, created endpoint to get all existing user paged

+++ server/src/Xyz.Core/Models/PageRequest.cs
@@ -2,8 +2,8 @@ namespace Xyz.Core.Models
 {
     public class PageRequest
     {
-        public long Index { get; set; }
-        public long Size { get; set; }
+        public int Index { get; set; }
+        public int Size { get; set; }
         public Sort Sort { get; set; } = default!;
     }
 }
